+++ src/main/java/com/example/quatro_to/service/FloorService.java
@@ -9,7 +9,7 @@ import java.util.Set;
 public interface FloorService {
 
     Floor save(Floor floor);
-    Floor findByNumber(Integer number);
+    Floor findByNumber(Long number);
     Floor findById(Long id);
     Floor update(Floor floor, Long id);
     void delete(Long id);
